docker/app/requirements.txt
pandas
minio
pyarrow
s3fs
psycopg2-binary
python-dotenv
boto3
numpy
